Modularized the blog posts
