+++ src/containers/app/routes/static/blog/blog.scss
@@ -1,125 +1,10 @@
 @import 'variables';
 
-$blog-post-border: $small-space;
-
 #blog {
-  #recent-posts {
-    margin-top: $large-space;
-    .lesser {
-      .blog-post {
-        @include desktop-up() {
-          height: 160px;
-          .excerpt {
-            display: none;
-          }
-        }
-      }
-    }
-  }
-  #posts-container {
-    @include desktop-up() {
-      margin-top: $large-space;
-    }
-  }
   #more-posts {
     text-align: center;
     button {
       margin: $medium-space 0px $large-space 0px;
-    }
-  }
-  .blog-post {
-    padding: $medium-space + $blog-post-border;
-    border-radius: $extra-border-radius;
-    margin-bottom: $medium-space;
-    background: $dark-gray;
-    box-shadow: 0px 10px 20px rgba($black, 0.1);
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    z-index: 0;
-    overflow: hidden;
-    @include desktop-up() {
-      height: 340px;
-    }
-    &:after {
-      position: absolute;
-      top: 0px;
-      left: 0px;
-      width: 100%;
-      height: 100%;
-      border: $blog-post-border solid rgba($white, 0.25);
-      z-index: -1;
-      content: '';
-    }
-    &.tutorials {
-      @extend .primary-diag-gradient;
-    }
-    &.roadmap {
-      @extend .success-diag-gradient;
-    }
-    &.announcements {
-      @extend .alert-diag-gradient;
-    }
-    &.unimportant {
-      @include desktop-up() {
-        background: $medium-gray;
-      }
-    }
-    .title {
-      color: $white;
-      font-size: 1.25rem;
-      margin: 0px;
-    }
-    .excerpt {
-      color: rgba($white, 0.75);
-      line-height: 1.75rem;
-      margin-bottom: 0px;
-      height: 115px;
-      overflow: hidden;
-    }
-    .metadata {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      height: 30px;
-      margin-top: $medium-space;
-      .details {
-        font-size: 0.7rem;
-        @include capped-text(1px);
-        color: rgba($white, 0.5);
-        .category {
-          color: rgba($black, 0.5);
-          &:hover {
-            color: $black;
-          }
-        }
-      }
-      .tags {
-        display: flex;
-        .tag {
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          background: $white;
-          border-radius: $border-radius;
-          height: 30px;
-          width: 30px;
-          padding: $small-space * 0.75;
-          margin-left: $small-space;
-          position: relative;
-          opacity: 0.5;
-          transition: opacity $anim-time $anim-func;
-          &:hover {
-            opacity: 1;
-          }
-          .repo-icon {
-            height: 100%;
-            width: 100%;
-            filter: contrast(0%) brightness(50%);
-          }
-        }
-      }
     }
   }
 }
